Remove py2 remnants (cast round() to int)

+++ mpmath/functions/zeta.py
@@ -876,7 +876,7 @@ def secondzeta_prime_term(ctx, s, a, **kwargs):
 
 def secondzeta_exp_term(ctx, s, a):
     if ctx.isnpint(s):
-        m = int(round(ctx.re(s)))
+        m = round(ctx.re(s))
         if not m & 1:
             return ctx.mpf('-0.25')**(-m//2)
     tol = ctx.eps
@@ -1025,7 +1025,7 @@ def secondzeta(ctx, s, a = 0.015, **kwargs):
     if ctx.isnpint(s-1):
         if abs(s-1) < tol*1000:
             return ctx.inf
-        m = int(round(ctx.re(s)))
+        m = round(ctx.re(s))
         if m & 1:
             return ctx.inf
         else:
